Fixes
- Removed uint64_t where the accurate value was needed (JS stores it as double internally, so it won't always return the same value after transforamtion).
- Fixed error locale.
- Fixed handling of JS::Value returns from native callbacks.

+++ foo_spider_monkey_panel/js_objects/fb_utils.h
@@ -43,7 +43,7 @@ public:
     std::optional<JSObject*> CreateProfiler( const std::string& name );
     std::optional<uint32_t> DoDragDrop( JsFbMetadbHandleList* handles, uint32_t okEffects );
     std::optional<std::nullptr_t> Exit();
-    std::optional<JSObject*> GetClipboardContents( uint64_t hWindow );
+    std::optional<JSObject*> GetClipboardContents( uint32_t hWindow );
     std::optional<std::string> GetDSPPresets();
     std::optional<JSObject*> GetFocusItem( bool force );
     std::optional<JSObject*> GetLibraryItems();
