+++ foo_spider_monkey_panel/js_objects/window.h
@@ -31,6 +31,8 @@ public:
     void RemoveHeapTracer();
 
 public: // methods
+    //std::optional<JS::Heap<JS::Value >> TestValue( uint32_t test );
+
     std::optional<std::nullptr_t> ClearInterval( uint32_t intervalId );
     std::optional<std::nullptr_t> ClearTimeout( uint32_t timeoutId );
     std::optional<JSObject*> CreatePopupMenu();
@@ -59,7 +61,7 @@ public: // methods
 public: // props
     std::optional<uint32_t> get_DlgCode();
     std::optional<uint32_t> get_Height();
-    std::optional<uint64_t> get_Id();
+    std::optional<uint32_t> get_Id();
     std::optional<uint32_t> get_InstanceType();
     std::optional<bool> get_IsTransparent();
     std::optional<bool> get_IsVisible();
